+++ Code/Libraries/Foundation/Common/Math/IntVector4.hpp
@@ -20,8 +20,8 @@ struct ZeroShared IntVector4
   }
   explicit IntVector4(int x, int y, int z, int w);
 
-  int& operator[](uint index);
-  int operator[](uint index) const;
+  int& operator[](size_t index);
+  int operator[](size_t index) const;
 
   // Unary Operators
   IntVector4 operator-(void) const;
@@ -83,7 +83,8 @@ struct ZeroShared IntVector4
   /// Set all of this vector's elements to 0.
   void ZeroOut(void);
 
-  union {
+  union
+  {
     struct
     {
       int x, y, z, w;
